+++ core/src/main/kotlin/no/elg/hex/util/SoundUtils.kt
@@ -7,6 +7,10 @@ fun playClick() {
   Hex.assets.clickSound?.play(Settings.volume)
 }
 
+fun playBadClick() {
+  Hex.assets.clickBadSound?.play(Settings.volume)
+}
+
 fun playMoney() {
   Hex.assets.coinsSound?.play(Settings.volume)
 }